rename uncategorized=>primitives
