+++ app/api/ada/adaTypes.js
@@ -8,7 +8,7 @@ import { RustModule } from './lib/cardanoCrypto/rustLoader';
 import type {
   DbTxIO
 } from './lib/storage/database/transactions/tables';
-import type { DbBlock, } from './lib/storage/database/uncategorized/tables';
+import type { DbBlock, } from './lib/storage/database/primitives/tables';
 import type {
   Address, Value, Addressing,
 } from './lib/storage/models/common/interfaces';
